+++ app/models/permitted_params.rb
@@ -232,7 +232,15 @@ class PermittedParams < Struct.new(:params, :user)
                                                       :planning_element_status_comment,
                                                       :planning_element_status_id,
                                                       :parent_id,
-                                                      :responsible_id
+                                                      :responsible_id,
+                                                      :custom_fields => [ #json
+                                                                          :id,
+                                                                          :value,
+                                                                          :custom_field => [ # xml
+                                                                                             :id,
+                                                                                             :value
+                                                                                           ]
+                                                                        ]
                                                     ],
                                :planning_element_type => [
                                                            :name,