+++ config/routes.rb
@@ -79,8 +79,13 @@ OpenProject::Application.routes.draw do
           get :available_projects, :on => :collection
         end
         resources :statuses, :only => [:index, :show]
+
+        member do
+          get :planning_element_custom_fields
+        end
       end
 
+      resources :custom_fields
 
       namespace :pagination, :as => 'paginate' do
         [:users,
@@ -93,6 +98,7 @@ OpenProject::Application.routes.draw do
           resources model, :only => [:index]
         end
       end
+
     end
   end